refactor: AllowedImageExtension filesystem 패키지로 이동
