--- backend/src/main/java/team/teamby/teambyteam/feed/domain/AllowedImageExtension.java
@@ -1,15 +1,0 @@
-package team.teamby.teambyteam.feed.domain;
-
-import java.util.Arrays;
-
-public enum AllowedImageExtension {
-
-    JPG, JPEG, PNG, WEBP, SVG, AI, BMP, AVIF;
-
-    public static boolean isNotContain(final String imageExtension) {
-        final String upperCase = imageExtension.toUpperCase();
-
-        return Arrays.stream(values())
-                .noneMatch(allowedImageExtension -> allowedImageExtension.name().equals(upperCase));
-    }
-}
